Updated with new objects
This update I didnt quite get to everything I wanted but i added two signs for understanding of the game and some new areas to interact with the box (and other boxes later on). Along with a new area for the point system.

I will be working on the score system and the timer on top.  Finally, making the forklift be able to pickup the objects.